requirements.txt
confusables==1.2.0
discord.py==2.3.1
rapidfuzz==1.4.1
pyflakes==2.2.0
